-code split on slow integration tests
