+++ integrationTests/relayers/slowTests/interface.go
@@ -1,5 +1,3 @@
-//go:build slow
-
 package slowTests
 
 import (
